% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pop_density.R
\name{apply_pop_density}
\alias{apply_pop_density}
\title{Apply population densityto a final RF prediction
(RF_pred +L1_pop)/RF_pred_ZS_sum We will need to change it
by using gdal_calc}
\usage{
apply_pop_density(
  input_poptables,
  censusmaskPathFileName,
  rfg.output.path.countries,
  nrpoc = NULL,
  rfg.countries.tag,
  quant = TRUE,
  minblocks = NULL,
  verbose = TRUE,
  log = FALSE
)
}
\arguments{
\item{input_poptables}{input poptablest}

\item{censusmaskPathFileName}{census mask Path FileName}

\item{rfg.output.path.countries}{path to load/save popfit objects}

\item{nrpoc}{number of cores}

\item{rfg.countries.tag}{proximity}

\item{quant}{proximity}

\item{minblocks}{minblocks}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
raster objects
}
\description{
Apply population densityto a final RF prediction
(RF_pred +L1_pop)/RF_pred_ZS_sum We will need to change it
by using gdal_calc
}
