% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_zs_covariates.R
\name{calculate_zonal_stats_covariates}
\alias{calculate_zonal_stats_covariates}
\title{calculate_zonal_stats_covariates calculate zonal stats}
\usage{
calculate_zonal_stats_covariates(
  x,
  y,
  input_poptables,
  save_zst = TRUE,
  nrpoc = NULL,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{x}{Input covariates list}

\item{y}{path to save results}

\item{input_poptables}{Input population}

\item{save_zst}{If TRUE save results}

\item{nrpoc}{NULL or Integer. Number of process to use}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
A data.frame compiled census_data
}
\description{
calculate_zonal_stats_covariates calculate zonal stats
}
\examples{
\dontrun{
calculate_zonal_stats_covariates( x, y, input_poptables)
}
}
