% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_zs_parallel.R
\name{calculate_zs_parallel}
\alias{calculate_zs_parallel}
\title{Function compute zonal statistics.
That is, cross-tabulate the values of a Raster* object based on a
"zones" RasterLayer. NA values are removed.
Function uses \code{\link[doParallel]{registerDoParallel}}
library to work with a big raster data.}
\usage{
calculate_zs_parallel(
  x,
  y,
  fun = "mean",
  cores = NULL,
  minblk = NULL,
  na.rm = TRUE,
  silent = TRUE
)
}
\arguments{
\item{x}{Raster* object}

\item{y}{RasterLayer object with codes representing zones}

\item{fun}{The function to be applied. Either as character: 'mean', 'min', 'max' and 'sum'}

\item{cores}{Integer. Number of cores for parallel calculation}

\item{minblk}{Integer. Minimum number of blocks}

\item{na.rm}{using na.rm = TRUE for missing data}

\item{silent}{If FALSE then the progress will be shown}
}
\value{
A data.frame with a value for each zone (unique value in zones)
}
\description{
Function compute zonal statistics.
That is, cross-tabulate the values of a Raster* object based on a
"zones" RasterLayer. NA values are removed.
Function uses \code{\link[doParallel]{registerDoParallel}}
library to work with a big raster data.
}
\examples{
\dontrun{
calculate_zs_parallel( x=rasterObj1, y=rasterObj2, cores=2, minblk=4  )
}
}
\author{
Maksym Bondarenko \href{mailto:mb4@soton.ac.uk}{mb4@soton.ac.uk} and
Chris Jochem \href{mailto:W.C.Jochem@soton.ac.uk}{W.C.Jochem@soton.ac.uk}
}
