% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_cov}
\alias{check_cov}
\title{Checking if covariates extend matches mastergrid.}
\usage{
check_cov(covariates, fix_cov, verbose = FALSE, log = FALSE)
}
\arguments{
\item{covariates}{list of covariates}

\item{fix_cov}{logical}

\item{verbose}{logical. Should report extra information on progress?}

\item{log}{logical. Should report on progress be saved in log file?}
}
\value{
logical if fix_raster is FALSE
}
\description{
Checking if covariates extend matches mastergrid.
}
