% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{check_raster_extend}
\alias{check_raster_extend}
\alias{change_raster_extend}
\title{Checking extent of two rasters}
\usage{
check_raster_extend(x, y)

change_raster_extend(srcfile, dstfile, verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{x}{path to raster}

\item{y}{path to raster}

\item{srcfile}{path to raster}

\item{dstfile}{path to raster}

\item{verbose}{logical. Should report extra information on progress?}

\item{overwrite}{logical to overwrite or not the output file}
}
\value{
logical

logical
}
\description{
Checking extent of two rasters

Changing extent of two rasters.
}
