% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_result.R
\name{check_result}
\alias{check_result}
\title{check_result comper input with output}
\usage{
check_result(
  input_poptables,
  censusmaskPathFileName,
  rfg.output.path.countries,
  nrpoc,
  rfg.countries.tag,
  minblocks = NULL,
  verbose = TRUE,
  log = FALSE
)
}
\arguments{
\item{input_poptables}{Input list of pop tables}

\item{censusmaskPathFileName}{path to census mask raster file}

\item{rfg.output.path.countries}{output dir}

\item{nrpoc}{number of processors}

\item{rfg.countries.tag}{tag of the project}

\item{minblocks}{minblocks.}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
Number of mean difference
}
\description{
check_result comper input with output
}
