% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{creat_raster_stack}
\alias{creat_raster_stack}
\title{creat_raster_stack Create a raster stack from all covariates from popfit and census_mask
and water_raster:}
\usage{
creat_raster_stack(covariates, popfit_final, census_mask, water_raster)
}
\arguments{
\item{covariates}{covariates list}

\item{popfit_final}{list of names used in the RF}

\item{census_mask}{raster of census mask}

\item{water_raster}{raster of water mask}
}
\value{
raster stack
}
\description{
creat_raster_stack Create a raster stack from all covariates from popfit and census_mask
and water_raster:
}
