% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates_list.R
\name{create_covariates_list}
\alias{create_covariates_list}
\title{create_covariates_list}
\usage{
create_covariates_list(
  input_covariates,
  input_mastergrid,
  input_watermask,
  input_px_area
)
}
\arguments{
\item{input_covariates}{Input list of covariates}

\item{input_mastergrid}{Input list to mastergrid}

\item{input_watermask}{Input list to watermask}

\item{input_px_area}{Input list to population}
}
\value{
A list of covariates
}
\description{
create_covariates_list
}
\examples{
\dontrun{
create_covariates_list( input_covariates, 
                        input_covariates, 
                        input_watermask, 
                        input_px_area  )
}
}
