% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates_list.R
\name{create_covariates_list_for_RF}
\alias{create_covariates_list_for_RF}
\title{create_covariates_list_for_RF}
\usage{
create_covariates_list_for_RF(covariates)
}
\arguments{
\item{covariates}{Input list of covariates}
}
\value{
A list of covariates
}
\description{
create_covariates_list_for_RF
}
\examples{
\dontrun{
create_covariates_list_for_RF( covariates )
}
}
