% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_dir}
\alias{create_dir}
\title{create_dir creating dir}
\usage{
create_dir(x, verbose)
}
\arguments{
\item{x}{name of the folder to be created}

\item{verbose}{If FALSE then the progress will be shown}
}
\description{
create_dir creating dir
}
