% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_dirs_for_prj}
\alias{create_dirs_for_prj}
\title{Create project directory based on tag of the countries.}
\usage{
create_dirs_for_prj(input.countries, output_dir)
}
\arguments{
\item{input.countries}{input list of the countries}

\item{output_dir}{path to project dir}
}
\value{
list of project directories
}
\description{
Create project directory based on tag of the countries.
}
