% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_file.R
\name{download_file}
\alias{download_file}
\title{Function to download file from ftp server}
\usage{
download_file(file_remote, dest_file, quiet, method = "auto")
}
\arguments{
\item{file_remote}{is a url to a remoute file}

\item{dest_file}{is a path where downloaded file will be stored}

\item{quiet}{If TRUE, suppress status messages (if any), and the progress bar.}

\item{method}{Method to be used for downloading files.
Current download methods are "internal", "wininet" (Windows only) "libcurl",
"wget" and "curl", and there is a value "auto"}
}
\description{
Function to download file from ftp server
}
