% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_aval_memory.R
\name{get_aval_memory}
\alias{get_aval_memory}
\title{get_aval_memory function will return avalible
of the system memory in GB
Tested on Windows 10}
\usage{
get_aval_memory()
}
\value{
numeric
}
\description{
get_aval_memory function will return avalible
of the system memory in GB
Tested on Windows 10
}
