% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_blocks_need.R
\name{get_blocks_need}
\alias{get_blocks_need}
\title{get_blocks_need function will return a number of blocks
sugesting for processing raster file. It will take into consideration
number of layers, cells, cores and avalible memory on computer
(not maximum memory but avalible)}
\usage{
get_blocks_need(x, cores, n = 1, number_type = "numeric")
}
\arguments{
\item{x}{raster}

\item{cores}{number of cores}

\item{n}{parameter to increase requrement of the raster}

\item{number_type}{Will be used to estimate requred memory}
}
\value{
integer
}
\description{
get_blocks_need function will return a number of blocks
sugesting for processing raster file. It will take into consideration
number of layers, cells, cores and avalible memory on computer
(not maximum memory but avalible)
}
\examples{
\dontrun{
get_blocks_need( x, cores=2, n=1 )
}
}
