% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covariates_list.R
\name{get_covariates_var_names}
\alias{get_covariates_var_names}
\title{get_covariates_var_names This function retrieves variable names
from the given covariates. Create a vector to hold the covariate
variable names:}
\usage{
get_covariates_var_names(x)
}
\arguments{
\item{x}{Input list of covariates}
}
\value{
A list of covariates
}
\description{
get_covariates_var_names This function retrieves variable names
from the given covariates. Create a vector to hold the covariate
variable names:
}
\examples{
\dontrun{
get_covariates_var_names( covariates )
}
}
