% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_system.R
\name{get_OS_system}
\alias{get_OS_system}
\title{get_OS_system function will return a string with OS
of the system
Tested on Windows 10}
\usage{
get_OS_system()
}
\value{
string
}
\description{
get_OS_system function will return a string with OS
of the system
Tested on Windows 10
}
