% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pop.R
\name{get_pop_census_all}
\alias{get_pop_census_all}
\title{get_pop_census_all load population data from input}
\usage{
get_pop_census_all(x)
}
\arguments{
\item{x}{input poptables}
}
\value{
A data.frame with the populations
}
\description{
get_pop_census_all load population data from input
}
