% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfit.R
\name{get_popfit}
\alias{get_popfit}
\title{get_popfit optimize the model}
\usage{
get_popfit(
  x_data,
  y_data,
  init_popfit = NULL,
  proximity = TRUE,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{x_data}{data for optimize}

\item{y_data}{data for optimize}

\item{init_popfit}{init_popfit}

\item{proximity}{proximity}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
randomForest objects
}
\description{
get_popfit optimize the model
}
