% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfit_final.R
\name{get_popfit_final}
\alias{get_popfit_final}
\title{get_popfit_final Another alternative is to use Quantile Regression Forests to generate
prediction intervals.  We'll fit a quantile regression using
the tuning parameters pulled from the popfit object.}
\usage{
get_popfit_final(
  x_data,
  y_data,
  popfit,
  popfit_fln,
  proximity = TRUE,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{x_data}{matrix or data frame of predictor variables}

\item{y_data}{response vector (factor for classification, numeric for regression)}

\item{popfit}{tuned randomForest objects}

\item{popfit_fln}{path to load/save popfit objects}

\item{proximity}{Should proximity measures be computed?}

\item{verbose}{logical. Should report extra information on progress?}

\item{log}{logical. Should report on progress be saved in log file?}
}
\value{
constructed popfit objects
}
\description{
get_popfit_final Another alternative is to use Quantile Regression Forests to generate
prediction intervals.  We'll fit a quantile regression using
the tuning parameters pulled from the popfit object.
}
