% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfit_final.R
\name{get_popfit_final_old}
\alias{get_popfit_final_old}
\title{get_popfit_final_old exit program without error}
\usage{
get_popfit_final_old(
  fset,
  only.names = FALSE,
  proximity = TRUE,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{fset}{list of the path to popfit.RData objects}

\item{only.names}{if true return only names}

\item{proximity}{Should proximity measures be computed?}

\item{verbose}{logical. Should report extra information on progress?}

\item{log}{logical. Should report on progress be saved in log file?}
}
\value{
previously constructed popfit.RData objects
}
\description{
get_popfit_final_old exit program without error
}
