% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfit_quant.R
\name{get_popfit_quant}
\alias{get_popfit_quant}
\title{get_popfit_quant Another alternative is to use Quantile Regression Forests to generate
prediction intervals.  We'll fit a quantile regression using
the tuning parameters pulled from the popfit object above:}
\usage{
get_popfit_quant(
  x_data,
  y_data,
  popfit,
  rfg.popfit.quant.RData,
  proximity = TRUE,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{x_data}{x data for randomForest}

\item{y_data}{y data for randomForest}

\item{popfit}{popfit objects}

\item{rfg.popfit.quant.RData}{path to load/save popfit objects}

\item{proximity}{proximity}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
constructed popfit objects
}
\description{
get_popfit_quant Another alternative is to use Quantile Regression Forests to generate
prediction intervals.  We'll fit a quantile regression using
the tuning parameters pulled from the popfit object above:
}
