% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfit_quant.R
\name{get_popfit_quant_old}
\alias{get_popfit_quant_old}
\title{get_popfit_quant_old exit program without error}
\usage{
get_popfit_quant_old(
  fset,
  only.names = FALSE,
  proximity = TRUE,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{fset}{list of the path to popfit.RData objects}

\item{only.names}{if true return only names}

\item{proximity}{proximity}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
previously constructed popfit.RData objects
}
\description{
get_popfit_quant_old exit program without error
}
