% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_check.R
\name{initial_check}
\alias{initial_check}
\title{stop_quietly exit program without error}
\usage{
initial_check(
  input_covariates,
  input_mastergrid,
  input_watermask,
  input_px_area,
  input_poptables,
  output_dir,
  nrpoc,
  minblocks
)
}
\arguments{
\item{input_covariates}{list of covariates}

\item{input_mastergrid}{list of mastergrid}

\item{input_watermask}{list of watermask}

\item{input_px_area}{list of px_area}

\item{input_poptables}{list of population table}

\item{output_dir}{name of the folder}

\item{nrpoc}{Number of process to use in script}

\item{minblocks}{min number of blocks for cluster}
}
\value{
stop execution of package
}
\description{
stop_quietly exit program without error
}
