% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pop.R
\name{load_pop}
\alias{load_pop}
\title{load_pop load population data from input}
\usage{
load_pop(icountry, input_poptables)
}
\arguments{
\item{icountry}{ISO of the country}

\item{input_poptables}{Path to the csv file with pop data}
}
\value{
A data.frame with the populations for each zone
}
\description{
load_pop load population data from input
}
