% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{log_info}
\alias{log_info}
\title{log_info save log into the file}
\usage{
log_info(prank, stext, verbose = FALSE, log = FALSE)
}
\arguments{
\item{prank}{type of log}

\item{stext}{Text to log}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
save log into the file
}
\description{
log_info save log into the file
}
