% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merg_covariates.R
\name{merg_covariates}
\alias{merg_covariates}
\title{merg_covariates merg covariates using gdal}
\usage{
merg_covariates(
  input.countries,
  covariates.var.names,
  covariates,
  rfg.countries.tag,
  rfg.countries.merged,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{input.countries}{list countries}

\item{covariates.var.names}{list of cov names}

\item{covariates}{detailes list of covariates}

\item{rfg.countries.tag}{tag for outputs}

\item{rfg.countries.merged}{directory to save merged rasters}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
A data.frame merged covariates
}
\description{
merg_covariates merg covariates using gdal
}
\examples{
\dontrun{
merg_covariates( covariates.var.names, 
                 covariates, 
                 rfg.countries.tag, 
                 rfg.countries.merged)
}
}
