% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popfit_init_tuning.R
\name{popfit_init_tuning}
\alias{popfit_init_tuning}
\title{Tuning of our randomForest population density regression}
\usage{
popfit_init_tuning(x, y, proximity = TRUE, verbose = FALSE, log = FALSE)
}
\arguments{
\item{x}{matrix or data frame of predictor variables}

\item{y}{response vector (factor for classification, numeric for regression)}

\item{proximity}{Should proximity measures be computed?}

\item{verbose}{logical. Should report extra information on progress?}

\item{log}{logical. Should report on progress be saved in log file?}
}
\value{
it returns a matrix whose first column contains the mtry values searched,
and the second column the corresponding OOB error
}
\description{
Tuning of our randomForest population density regression
}
