% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popRF.R
\name{popRF}
\alias{popRF}
\title{popRF Random Forests population modelling scripts
Southampton University
WordlPop Spatial Data Infrastructure
https://sdi.worldpop.org}
\usage{
popRF(
  pop,
  cov,
  mastergrid,
  watermask,
  px_area,
  output_dir,
  npoc = NULL,
  minblocks = NULL,
  quant = TRUE,
  proximity = TRUE,
  fset = NULL,
  fset_incl = FALSE,
  fset_cutoff = 20,
  fix_cov = FALSE,
  check_result = FALSE,
  verbose = TRUE,
  log = FALSE
)
}
\arguments{
\item{pop}{Input list to population}

\item{cov}{Input list of covariates}

\item{mastergrid}{Input list to mastergrid}

\item{watermask}{Input list to watermask}

\item{px_area}{Input list to population}

\item{output_dir}{Path to the folder to save the outputs}

\item{npoc}{NULL or Integer. Number of process to use}

\item{minblocks}{Integer. if minblocks  <- NULL then minblocks for cluster
prediction parallesation will be calculated based on aval memory}

\item{quant}{If FALSE then quant will not be calculated}

\item{proximity}{Should proximity measures be computed?}

\item{fset}{Declare if we are using a fixed set in this modeling,
i.e. are we parameterizing, in part or in full, this RF model run upon another
country's(ies') RF model object.}

\item{fset_incl}{If TRUE  RF popfit will not be combined with RF model
run upon another country's(ies') RF model object}

\item{fset_cutoff}{Default 20}

\item{fix_cov}{logical. Should try to fix extend of covariates if they do not match mastergrid?}

\item{check_result}{If TRUE then the results will be compare with input}

\item{verbose}{logical. Should report extra information on progress?}

\item{log}{logical. Should report on progress be saved in log file?}
}
\value{
raster object of population
}
\description{
Authors & Maintainer of the script
Maksym Bondarenko \href{mailto:mb4@soton.ac.uk}{mb4@soton.ac.uk}
Jeremiah J. Nieves \href{mailto:jeremiah.j.nieves@outlook.com}{jeremiah.j.nieves@outlook.com}
David Kerr \href{mailto:dk2n16@soton.ac.uk}{dk2n16@soton.ac.uk}
Alessandro Sorichetta \href{mailto:as1v13@soton.ac.uk}{as1v13@soton.ac.uk}
}
\details{
Random Forest (RF)-based dasymetric mapping approach developed
by Stevens et al. (2015)*
}
\note{
\itemize{
\item Stevens, F. R., Gaughan, A. E., Linard, C. & Tatem, A. J.
Disaggregating Census Data for Population Mapping Using Random Forests
with Remotely-Sensed and Ancillary Data. PLoS ONE 10, e0107042 (2015).
http://journals.plos.org/plosone/article?id=10.1371/journal.pone.0107042
}
}
\examples{
\dontrun{

pop_table <- list("Nepal"="/user/npl_population.csv")

input_cov <- list(
                 "Nepal"=list("dst011"="/user/dst011.tif",
                              "dst040"="/user/dst040.tif",
                              "dst130"="/user/dst130.tif"
                              )  
                 )

input_mastergrid <- list("Nepal"="/user/mastergrid.tif")
input_watermask <- list("Nepal"="/user/watermask.tif")
input_px_area <- list("Nepal"="/user/px_area.tif")

popRF(pop=pop_table, 
      cov=input_cov, 
      mastergrid=input_mastergrid, 
      watermask=input_watermask, 
      px_area=input_px_area, 
      output_dir="/user/output", 
      cores=4 
      )
}
}
