% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popRFdemo.R
\name{popRFdemo}
\alias{popRFdemo}
\title{Function to demo the popFR package using WorldPop input data}
\usage{
popRFdemo(project_dir, country="NPL", cores=4, quant=TRUE, ftp=TRUE, 
verbose=TRUE, log=TRUE)
}
\arguments{
\item{project_dir}{Path to the folder to save the outputs}

\item{country}{ISO of the country. Default one is NPL (Nepal)}

\item{cores}{is a integer. Number of cores to use when executing the function,
which defaults to 4. If set to 0 or NULL max number of cores will be
used based on as many processors as the hardware and RAM allow.}

\item{quant}{If FALSE then quant will not be calculated}

\item{ftp}{is logical. TRUE or FALSE: flag indicating whether
\href{ftp://ftp.worldpop.org}{FTP} or \href{https://data.worldpop.org}{HTTPS}
WorldPop data hub server will be used. Default is \code{ftp} = TRUE.}

\item{verbose}{is logical. TRUE or FALSE: flag indicating whether to print
intermediate output from the function on the console, which might be
helpful for model debugging. Default is \code{verbose} = TRUE.}

\item{log}{is logical. TRUE or FALSE: flag indicating whether to print intermediate
output from the function on the log.txt file.
Default is \code{log} = FALSE.}
}
\value{
A data.frame with a value for each zone (unique value in zones)
}
\description{
Function to demo the popFR package using WorldPop input data
}
\examples{
\dontrun{
popRFdemo( project_dir="/home/user/demo",
           country="NPL", 
           cores=4
         )
} 
}
