% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress_message.R
\name{progress_message}
\alias{progress_message}
\title{progress_message function will return progress}
\usage{
progress_message(x, max = 100, label = NULL)
}
\arguments{
\item{x}{integer}

\item{max}{maximum for progress bvar}

\item{label}{additional text for progress  bar}
}
\value{
character
}
\description{
Tested on Windows 10
}
\examples{
\dontrun{
progress_message( x=10, max = 200, label="Progress message" )
}
}
