% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_parallel.R
\name{rasterize_parallel}
\alias{rasterize_parallel}
\title{Function will transfer values associated with 'object' type spatial
data (data.frame) to raster cells. Function is using paralle library to work
with a big raster data. The raster file will be splited to blocks and
processed per block.}
\usage{
rasterize_parallel(x, df, cores=NULL, minblk=NULL, NAflag=NULL, 
                          datatype=NULL, filename=rasterTmpFile(), 
                          overwrite=TRUE, silent=TRUE)
}
\arguments{
\item{x}{Raster* object.}

\item{df}{data.frame of points.}

\item{cores}{is a integer. Number of cores to use when executing the
function in paralle.}

\item{minblk}{Integer. Minimum number of blocks.}

\item{NAflag}{NO data value will be used for a new raster.}

\item{datatype}{Type of raster. Available are
INT1S/INT2S/INT4S/FLT4S/LOG1S/INT1U/INT2U/INT4U/FLT8S.}

\item{filename}{the name of the raster file.}

\item{overwrite}{Overwrite existing file.}

\item{silent}{is logical. TRUE or FALSE: flag indicating whether to print
intermediate output from the function on the console, which might be
helpful for model debugging. If FALSE then the progress will be shown.
Default is \code{silent} = TRUE.}
}
\value{
Raster* object
}
\description{
Function will transfer values associated with 'object' type spatial
data (data.frame) to raster cells. Function is using paralle library to work
with a big raster data. The raster file will be splited to blocks and
processed per block.
}
\examples{
\dontrun{
rasterize_parallel(x=rasterObj, df=df, cores=2,minblk=4,NAflag=-99999,datatype='INT1U' )
}
}
\author{
Maksym Bondarenko \href{mailto:mb4@soton.ac.uk}{mb4@soton.ac.uk} and
Chris Jochem \href{mailto:W.C.Jochem@soton.ac.uk}{W.C.Jochem@soton.ac.uk}
}
