% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterize_parallel.R
\name{rasterize_parallel_start}
\alias{rasterize_parallel_start}
\title{rasterize_parallel_start function will return a string with OS}
\usage{
rasterize_parallel_start(
  x,
  df,
  minblk,
  NAflag,
  datatype,
  filename,
  overwrite = TRUE,
  silent = TRUE
)
}
\arguments{
\item{x}{Raster* object}

\item{df}{data.frame of points}

\item{minblk}{Integer. Minimum number of blocks}

\item{NAflag}{NO data value will be used for a new raster}

\item{datatype}{Type of raster. Avalible are INT1S/INT2S/INT4S/FLT4S/LOG1S/INT1U/INT2U/INT4U/FLT8S}

\item{filename}{File of a new raster file.}

\item{overwrite}{Overwrite existing file}

\item{silent}{If FALSE then the progress will be shown}
}
\value{
Raster* object
}
\description{
rasterize_parallel_start function will return a string with OS
}
