% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_prediction.R
\name{rf_prediction}
\alias{rf_prediction}
\title{rf_prediction Predict for gridded covariates}
\usage{
rf_prediction(
  covariates,
  census_mask,
  water_raster,
  popfit_final,
  popfit_quant,
  outdir,
  tag,
  quant = TRUE,
  verbose = FALSE,
  log = FALSE
)
}
\arguments{
\item{covariates}{covariates list}

\item{census_mask}{census_mask}

\item{water_raster}{water_raster}

\item{popfit_final}{popfit final objects}

\item{popfit_quant}{popfit quant objects}

\item{outdir}{path to load/save popfit objects}

\item{tag}{proximity}

\item{quant}{proximity}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
raster objects
}
\description{
rf_prediction Predict for gridded covariates
}
