% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_prediction_parallel.R
\name{rf_prediction_parallel}
\alias{rf_prediction_parallel}
\title{rf_prediction_parallel Predict for gridded covariates}
\usage{
rf_prediction_parallel(
  covariates,
  census_mask,
  water_raster,
  popfit_final,
  popfit_quant,
  outdir,
  nrpoc,
  tag,
  quant = TRUE,
  minblocks = NULL,
  verbose = TRUE,
  log = FALSE
)
}
\arguments{
\item{covariates}{popfit quant objects}

\item{census_mask}{census_mask}

\item{water_raster}{water_raster}

\item{popfit_final}{covariates list}

\item{popfit_quant}{popfit final objects}

\item{outdir}{path to load/save popfit objects}

\item{nrpoc}{path to load/save popfit objects}

\item{tag}{proximity}

\item{quant}{proximity}

\item{minblocks}{proximity}

\item{verbose}{If FALSE then the progress will be shown}

\item{log}{If FALSE then the progress will be shown}
}
\value{
raster objects
}
\description{
rf_prediction_parallel Predict for gridded covariates
}
