% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{specify_decimal}
\alias{specify_decimal}
\title{This function allows for precision decimal formatting.}
\usage{
specify_decimal(x, k)
}
\arguments{
\item{x}{number}

\item{k}{format}
}
\value{
number
}
\description{
This function allows for precision decimal formatting.
}
