% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{stop_quietly}
\alias{stop_quietly}
\title{stop_quietly exit program without error}
\usage{
stop_quietly()
}
\value{
stop execution of package
}
\description{
stop_quietly exit program without error
}
