% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tmDiff}
\alias{tmDiff}
\title{tmDiff return a time difference in human  reading format}
\usage{
tmDiff(start, end, frm = "hms")
}
\arguments{
\item{start}{time when started}

\item{end}{time when endted}

\item{frm}{format of return}
}
\description{
tmDiff return a time difference in human  reading format
}
