% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{transY}
\alias{transY}
\title{transY The default is to log transform the x argument:}
\usage{
transY(x, inverse = FALSE)
}
\arguments{
\item{x}{vector}

\item{inverse}{if TRUE inverse}
}
\value{
A data.frame merged covariates
}
\description{
transY The default is to log transform the x argument:
}
