% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popRF.R
\name{popRF}
\alias{popRF}
\title{Disaggregating Census Data for Population Mapping Using Random Forests
with Remotely-Sensed and Ancillary Data.}
\usage{
popRF(pop, cov, mastergrid, watermask, px_area, output_dir, cores=0, 
minblocks=NULL, quant=TRUE, proximity=TRUE, fset=NULL, fset_incl=FALSE, 
fset_cutoff=20, fix_cov=FALSE, check_result=TRUE, verbose=TRUE, log=FALSE)
}
\arguments{
\item{pop}{the name of the file which the administrative ID and the population
values are to be read from. The file should contain two columns
comma-separated with the value of administrative ID and population
without columns names. If it does not contain an absolute path, the
file name is relative to the current working directory.}

\item{cov}{List contains a list of elements. Each element of a list is another
list object with a given name of the country, the element of this list
is the input covariates with the name of the covariates and the path to
them.
Example:\if{html}{\out{<div class="r">}}\preformatted{list(
    "NPL"=list(
               "covariate1" = "covariate1.tif",
               "covariate2" = "covariate2.tif"
              )  
   )
}\if{html}{\out{</div>}}\preformatted{## $NPL
## $NPL$covariate1
## [1] "covariate1.tif"
## 
## $NPL$covariate2
## [1] "covariate2.tif"
}}

\item{mastergrid}{List with each element of a list is
another object with a given name of the country, the element of
this list is the input mastergrid with the path to the raster file.
Example:\preformatted{list(
    "NPL" = "npl_mastergrid.tif"
   )
}}

\item{watermask}{List with each element of a list is
another object with a given name of the country, the element of
this list is the input watermask with the path to the raster file.
Example:\preformatted{list(
    "NPL" = "npl_watermask.tif"
   )
}}

\item{px_area}{List with each element of a list is
another object with a given name of the country, the element of
this list is the input px_area with the path to the raster file.
Example:\if{html}{\out{<div class="r">}}\preformatted{list(
    "NPL" = "npl_px_area.tif"
   )
}\if{html}{\out{</div>}}\preformatted{## $NPL
## [1] "npl_px_area.tif"
}}

\item{output_dir}{path for the location of the output.
Default is the temp directory.}

\item{cores}{is a integer. Number of cores to use when executing the function.
If set to 0 \code{(max_number_of_cores - 1)}  will be used based on as
many processors as the hardware and RAM allow.
Default is \code{cores} = 0.}

\item{minblocks}{Integer. if \code{minblocks} is NULL then \code{minblocks}
for cluster prediction parallesation will be calculated based on
available memory.}

\item{quant}{is logical. TRUE or FALSE: flag indicating whether to produce
the quantile regression forests to generate prediction intervals.
Default is \code{quant} = TRUE.}

\item{proximity}{is logical. TRUE or FALSE: flag indicating whether proximity
measures among the rows be computed? Default is \code{proximity} = TRUE.
See \code{\link[randomForest]{randomForest}} for more details.}

\item{fset}{Declare if we are using a fixed set in this modeling, i.e. are we
parameterizing, in part or in full, this RF model run upon another
country's(ies') RF model object.}

\item{fset_incl}{is logical. TRUE or FALSE: flag indicating whether RF model object
will or will not be combined with RF model run upon another
country's(ies') RF model object.}

\item{fset_cutoff}{is a integer. This parameter only used if \code{fset_incl}
is TRUE. If the country has less then \code{fset_cutoff} idmin units
then RF popfit will not be combined with RF model run upon another
country's(ies') RF model object. Default is \code{fset_cutoff} = 20.}

\item{fix_cov}{is logical. TRUE or FALSE: flag indicating whether the raster
extend of the covariates will be fixed if the extend does not match
mastergrid. \code{fix_cov} = FALSE.}

\item{check_result}{is logical. TRUE or FALSE: flag indicating whether the
results will be compare with input data.
Default is \code{check_result} = TRUE.}

\item{verbose}{is logical. TRUE or FALSE: flag indicating whether to print
intermediate output from the function on the console, which might be
helpful for model debugging. Default is \code{verbose} = TRUE.}

\item{log}{is logical. TRUE or FALSE: flag indicating whether to print intermediate
output from the function on the log.txt file.
Default is \code{log} = FALSE.}
}
\value{
Raster* object of gridded population.
}
\description{
Disaggregating Census Data for Population Mapping Using Random Forests
with Remotely-Sensed and Ancillary Data.
}
\details{
This function producing gridded population density estimates using
a Random Forest model as described in
\href{https://doi.org/10.1371/journal.pone.0107042}{Stevens, et al. (2015)}.
The population density and covariate aggregation values for each
census unit are then used to create a
\href{https://doi.org/10.1023/A:1010933404324}{Random Forest model}
to predict log population density. Random Forest models are an
ensemble, nonparametric modeling approach that grows a "forest" of
individual classification or regression trees and improves upon
bagging by using the best of a random selection of predictors at
each node in each tree. Model estimation, fitting and an estimated
prediction weighting layer is used to dasymetrically redistribute
the census counts.
}
\examples{
\dontrun{

library("popRF")

pop_table <- list("NPL"="/user/npl_population.csv")

input_cov <- list(
                 "NPL"=list(
                            "cov1" = "covariate1.tif",
                            "cov2" = "covariate2.tif"))
                            
                 
input_mastergrid <- list("NPL" = "npl_mastergrid.tif")
input_watermask  <- list("NPL" = "npl_watermask.tif")
input_px_area    <- list("NPL" = "npl_px_area.tif")

res <- popRF(pop=pop_table, 
             cov=input_cov, 
             mastergrid=input_mastergrid, 
             watermask=input_watermask, 
             px_area=input_px_area, 
             output_dir="/user/output", 
             cores=4) 
 
# Plot populataion raster 
plot(res$pop) 

# Plot Error via Trees     
plot(res$popfit)
            
}
}
\references{
Stevens, F. R., Gaughan, A. E., Linard, C. & Tatem,
A. J. Disaggregating Census Data for Population Mapping Using Random
Forests with Remotely-Sensed and Ancillary Data. PLoS ONE 10, e0107042
(2015). \url{https://doi.org/10.1371/journal.pone.0107042}
}
\author{
Maksym Bondarenko \href{mailto:mb4@soton.ac.uk}{mb4@soton.ac.uk},
Jeremiah J. Nieves \href{mailto:jeremiah.j.nieves@outlook.com}{jeremiah.j.nieves@outlook.com},
Forrest R. Stevens \href{mailto:forrest.stevens@louisville.edu}{forrest.stevens@louisville.edu},
Andrea E. Gaughan \href{mailto:ae.gaughan@louisville.edu}{ae.gaughan@louisville.edu},
David Kerr \href{mailto:dk2n16@soton.ac.uk}{dk2n16@soton.ac.uk},
Chris Jochem \href{mailto:W.C.Jochem@soton.ac.uk}{W.C.Jochem@soton.ac.uk} and
Alessandro Sorichetta \href{mailto:as1v13@soton.ac.uk}{as1v13@soton.ac.uk}
}
