% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popRFdemo.R
\name{popRFdemo}
\alias{popRFdemo}
\title{Function to demo the popFR package using WorldPop inout data}
\usage{
popRFdemo(
  project_dir,
  country = "NPL",
  npoc = 4,
  quant = TRUE,
  ftp = TRUE,
  verbose = TRUE,
  log = TRUE
)
}
\arguments{
\item{project_dir}{Path to the folder to save the outputs}

\item{country}{ISO of the country in demo. Default one is NPL (Nepal)}

\item{npoc}{NULL or Integer. Number of process to use}

\item{quant}{If FALSE then quant will not be calculated}

\item{ftp}{Method to be used for downloading files. If TRUE FTP will be used otherwise https}

\item{verbose}{logical. Should report extra information on progress?}

\item{log}{logical. Should report on progress be saved in log file?}
}
\value{
A data.frame with a value for each zone (unique value in zones)
}
\description{
Function to demo the popFR package using WorldPop inout data
}
\examples{
\dontrun{
popRFdemo( project_dir="/home/user/demo",
           country="NPL", 
           npoc=4
         )
} 
}
